endpoints ok - refatorado nome do attr company pra controlNumber pra nao ter confusao

+++ src/main/kotlin/br/com/vipsolutions/connect/model/ws/ActionWs.kt
@@ -8,7 +8,7 @@ import br.com.vipsolutions.connect.model.CompanyInfo
  */
 class ActionWs(
     val action: String,
-    val controlNumber: Int,
+    val controlNumber: Long,
     val instanceId: Long,
     val qrCode: QrCode?,
     val companyInfo: CompanyInfo?
